Update target frameworks - Update target frameworks to support .NET7 - Includes settings AnalysisMode to recommended and the resulting changes
