+++ src/HidApi.Net.Tester/Verifier/Linux.cs
@@ -1,11 +1,11 @@
 namespace HidApi.Net.Tester;
 
-public class Linux : HidApiVerifier
+public class Linux : IHidApiVerifier
 {
     public bool VerifyCharPointer()
     {
         var version = Hid.VersionString();
-        if (version.Split('.').Length == 3 && version.StartsWith("0."))
+        if (version.Split('.').Length == 3 && version.StartsWith("0.", StringComparison.InvariantCulture))
             return true;
 
         Console.WriteLine("Error: Could not verify HIDAPI. String seems to have incorrect format.");
