+++ src/HidApi.Net/Internal/NullTerminatedString.cs
@@ -1,3 +1,5 @@
+using System.Text;
+
 namespace HidApi;
 
 internal readonly ref struct NullTerminatedString
@@ -11,9 +13,27 @@ internal readonly ref struct NullTerminatedString
         data = ReadOnlySpan<byte>.Empty;
     }
 
-    internal NullTerminatedString(ref byte[] str)
+    private NullTerminatedString(ReadOnlySpan<byte> data)
     {
-        data = str;
+        this.data = data;
+    }
+
+    internal static NullTerminatedString WithUnicode(string str)
+    {
+        var src = Encoding.Unicode.GetBytes(str);
+        var dest = new byte[src.Length + sizeof(ushort)];
+        Array.Copy(src, dest, src.Length);
+
+        return new NullTerminatedString(dest);
+    }
+
+    internal static NullTerminatedString WithUtf32(string str)
+    {
+        var src = Encoding.UTF32.GetBytes(str);
+        var dest = new byte[src.Length + sizeof(uint)];
+        Array.Copy(src, dest, src.Length);
+
+        return new NullTerminatedString(dest);
     }
 
     public ref readonly byte GetPinnableReference() => ref data.GetPinnableReference();
